Add files via upload
Customer Churn Prediction:

📌 Problem Definition:

                The `Customer Churn table` contains information on all `7,043` customers from a `Telecommunications company` in California in Q2 2022.Each record represents `one customer`, and contains details about their `demographics`, `location`, `tenure`, `subscription services`, `status for the quarter` `(joined, stayed, or churned)`, and more!
The `Zip Code Population` table contains complimentary information on the estimated populations for the California zip codes in the Customer Churn table. We need to `predict` whether the customer will `churn`, `stay` or `join` the company based on the parameters of the dataset.

📓 Overview:

| Machine Learning Models Applied | Accuracy |
| -------------------------------------- | ---------- |
| Random Forest                                |   78.11%  |
| Logistic Regression                          |   78.28%  |
| Naive Bayes Gaussian                      |   36.77%  |
| Decision Tree                                   |   77.29%  |
| XGB_Classifier                                  |   80.86%  |

👉 Application:

The ability to predict churn before it happens allows businesses to take proactive actions to keep existing customers from churning. This could look like: 
```
  Customer success teams reaching out to those high-risk customers to provide support or to gauge 
  what needs may not be being met.
```
The advantage of calculating a company's churn rate is that it provides clarity on how well the business is retaining customers, which is a reflection on the quality of the service the business is providing, as well as its usefulness.